requirements.txt
Flask==1.0
Flask-Cors==3.0.2
Flask-HTTPAuth==1.0.0
Flask-MySQL==1.4.0
Flask-RESTful==0.3.5
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.11.3
jsondatabase==0.1.7
jsonschema==2.6.0
MarkupSafe==0.23
Werkzeug==0.16.0
flask-blueprint==1.2.7
python-binary-memcached==0.28.0
Flask-Caching==1.4.0
pylibmc==1.5.2
mysql-connector-python==8.0.23